test(playwright): fix code coverage on all tests
