+++ test/src/example.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from '@playwright/test';
-import { test } from '../baseTest';
+import { test } from 'test/fixtures/baseTest';
 
 test('Click home link', async ({ page }) => {
   await page.goto('/');
